added console logs & cover page generator

+++ src/main.ts
@@ -1,43 +1,90 @@
-import { captionImage, cropAndResizeImage } from "./generator";
+import { captionImage, cropAndResizeImage, createCoverPage } from "./generator";
 import { getImageInfoPack, downloadImage, imageWriter } from "./image_puller";
-
+import dotenv from "dotenv";
 import { 
     createDataFolders, 
     getConfigLines, 
     getRandomNumber, 
     getRandomThemeAndRemove, 
     randomNumbersArray, 
-    deleteFolder 
+    deleteFolder,
+    existData
 } from "./util";
+import { bgRedBright, bgYellowBright, bold, options, bgGreenBright, bgBlueBright } from "colorette";
 
 // create default folders
 createDataFolders();
+dotenv.config();
+options.enabled = (process.env.TERMINAL_COLORS == "true") ? true : false;
 
-// Proof of concept
+if (!existData(process.env.NAME_LIST||"names.txt")) {
+    console.log(`${bgRedBright(`${bold(" ERROR ")}`)} There is nothing in the ${bold("name")} file`);
+    process.exit();
+} 
+
+if (!existData(process.env.THEMES_LIST||"themes.txt")) {
+    console.log(`${bgRedBright(`${bold(" ERROR ")}`)} There is nothing in the ${bold("theme")} file`);
+    process.exit();
+} 
+
 const theme: string = getRandomThemeAndRemove();
 const info_pack = getImageInfoPack(theme);
 
-const list_name: string[] = getConfigLines("names.txt");
+const list_name: string[] = getConfigLines(process.env.NAME_LIST||"names.txt");
 
 info_pack.then( async (res) => {
-    const random_array: number[] = randomNumbersArray(0, res.length, list_name.length);
-    
+
+    let coverpage_download = await downloadImage(res[getRandomNumber(0, 300)].image);
+    console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Downloading ${bold(theme)} cover page.`);
+
+    // Create cover page image & crop & resize
+    while(true) {
+        // if something wrong with the download, re-download
+        if (coverpage_download.statusCode != 200) {
+            coverpage_download = await downloadImage(res[getRandomNumber(0, 300)].image);
+            console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Re-downloading ${bold(theme)} cover page.`);
+        } else break;
+    }
+    console.log(`${bgGreenBright(`${bold(" SUCCESSFUL ")}`)} Downloaded ${bold("COVER_PAGE.JPG")} image from ${bold(theme)} theme.`);
+    console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Writing ${bold(theme)} cover page.`);
+    imageWriter(coverpage_download.data, theme, "COVER_PAGE");
+
+    console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Cropping & Resizing ${bold(theme)} cover page.`);
+    cropAndResizeImage("COVER_PAGE", theme);
+
+    console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Creating ${bold(theme)} cover page.`);
+    createCoverPage(theme);
+
+    console.log(`${bgGreenBright(`${bold(" STARTING ")}`)} Generating random numbers set.`);
+    const random_array: number[] = randomNumbersArray(0, 100, list_name.length);
     random_array.forEach( async (random_number: number, i: number) => {
         let download = await downloadImage(res[random_number].image);
+        console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Downloading ${bold(list_name[i])} image from ${bold(theme)} theme.`);
 
         while(true) {
             // if something wrong with the download, re-download
-            if (download.statusCode != 200)
+            if (download.statusCode != 200) {
                 download = await downloadImage(res[getRandomNumber(0, 100)].image);
-            else break;
+                console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Re-downloading ${bold(list_name[i])} image from ${bold(theme)} theme.`);
+            } else break;
         }
 
         // Save the image
+        console.log(`${bgGreenBright(`${bold(" SUCCESSFUL ")}`)} Downloaded ${bold(list_name[i])} image from ${bold(theme)} theme.`);
+        
+        console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Writing ${bold(list_name[i])} image from ${bold(theme)} theme.`);
         imageWriter(download.data, theme, list_name[i]);
+        
+        console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Cropping & Resizing ${bold(list_name[i])} image from ${bold(theme)} theme.`);
         cropAndResizeImage(list_name[i], theme);
+        
+        console.log(`${bgYellowBright(`${bold(" TRYING ")}`)} Putting captions ${bold(list_name[i])} image from ${bold(theme)} theme.`);
         captionImage(list_name[i], theme, list_name[i]);
+
+        // deleting (finishing process)
+        console.log(`${bgBlueBright(`${bold(" FINISHED ")}`)} Removing ${bold("input")} & ${bold("cropped")} folders.`);
         deleteFolder(`/input/images/${theme}`);
         deleteFolder(`/cropped/images/${theme}`);
-        
     });
+
 });
